move to zho and fix revision
