+++ mteb/tasks/Classification/zho/YueOpenriceReviewClassification.py
@@ -11,12 +11,12 @@ class YueOpenriceReviewClassification(AbsTaskClassification):
         reference="https://github.com/Christainx/Dataset_Cantonese_Openrice",
         dataset={
             "path": "izhx/yue-openrice-review",
-            "revision": "b8b9a3d0e40bc0674945a0edb0bc549b554d6f20",
+            "revision": "1300d045cf983bac23faadf3aa12a619624769da",
         },
         type="Classification",
         category="s2s",
         eval_splits=["test"],
-        eval_langs={"yue": ["yue-Hant"]},
+        eval_langs=["yue-Hant"],
         main_score="accuracy",
         date=("2019-01-01", "2019-05-01"),
         form=["spoken"],
