+++ results/intfloat__multilingual-e5-small/YueOpenriceReviewClassification.json
@@ -1,11 +1,11 @@
 {
-  "dataset_revision": "b8b9a3d0e40bc0674945a0edb0bc549b554d6f20",
+  "dataset_revision": "1300d045cf983bac23faadf3aa12a619624769da",
   "mteb_dataset_name": "YueOpenriceReviewClassification",
   "mteb_version": "1.6.3",
   "test": {
     "accuracy": 0.2950170426878754,
     "accuracy_stderr": 0.01689226480107389,
-    "evaluation_time": 16.75,
+    "evaluation_time": 16.23,
     "f1": 0.25676375263250273,
     "f1_stderr": 0.009293715552198299,
     "main_score": 0.2950170426878754
